+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Extract MARC Relators data for compilers (`marcrel:com`)
 * Extract MARC Relators data for contributors (`marcrel:ctb`)
+* Extract book Series from `pgterms:marc440`